+++ src/image-renderer.cpp
@@ -16,7 +16,10 @@ You should have received a copy of the GNU General Public License
 along with image-renderer.  If not, see <https://www.gnu.org/licenses/>. */
 
 #include "renderer.h"
+#include "parser.h"
+#include "texture.h"
 #include "argh.h"
+
 #undef NDEBUG
 #include <cassert>
 #include <sstream>
@@ -24,79 +27,115 @@ along with image-renderer.  If not, see <https://www.gnu.org/licenses/>. */
 #include <iostream>
 #include <iomanip>
 #include <vector>
+#include <unordered_map>
 
 #define USAGE \
+	programName << ": a C++ tool to generate photo-realistic images." << endl << endl << \
 	"Usage: " << endl << \
 	programName << " -h|--help" << endl << \
+	programName << " render [options] <inputfile>" << endl << \
 	programName << " demo [options]" << endl << \
-	programName << " pfm2ldr [options] <format> <inputfile> <outputfile>" << \
-	programName << " stack [options] <inputfiles>" \
-	" <input-pfm-file> <output-file> " << endl
-
-#define RUN_HELP \
-	"Run '" << programName << " <action-name> -h' for all supported options." << endl
+	programName << " pfm2ldr [options] <inputfile>" << endl << \
+	programName << " stack [options] <inputfiles>" << endl << endl << \
+	"Run '" << programName << " <action-name> -h|--help' for all supported options." << endl
 
 #define HELP_PFM2LDR \
-	"Available actions:" << endl << \
-	"-h, --help: print this message." << endl << endl << \
-	"pfm2ldr: convert a pfm image to a ldr image in the desired format." << endl << \
-	"	Common options:" << endl << \
-	"		-a <value>, --afactor=<value>			Normalization factor." << endl << \
-	"		-g <value>, --gamma=<value>			Gamma factor." << endl << endl << \
-	"	Supported formats and related options:" << endl << \
-	"		bmp" << endl << \
-	"		gif" << endl << \
-	"		jpeg 	-q <value>, --quality=<value>		Compression quality (0-95, 0 means default)." << endl << \
-	"		png	-p, --palette				Whether to use palette or not."<< endl << \
-	"			-c <value>, --compression=<value>	Compression level (0-9)." << endl << \
-	"		tiff" << endl << \
-	"		webp	-q <value>, --quality=<value>		Compression quality (0-100)." << endl
+	"pfm2ldr: convert a pfm image to a ldr image in the desired format." << endl << endl << \
+	"Usage:	" << programName << " pfm2ldr [options] <inputfile>" << endl << endl << \
+	"General options:" << endl << \
+	"	-h, --help					Print this message." << endl << \
+	"	-f <format>, --format=<format>			Format of the output ldr image (default 'png'). See below for a list of supported formats." << endl << \
+	"	-o <string>, --outfile=<string>			Filename of output image (default input filename with format-dependent extension)." << endl << \
+	"	-l <value>, --luminosity=<value>		Total luminosity normalization factor (by default, it is calculated from the input pfm)." << endl << \
+	"	-a <value>, --afactor=<value>			Normalization coefficient (default 0.3)." << endl << \
+	"	-g <value>, --gamma=<value>			Gamma factor (default 1)." << endl << endl << \
+	"Supported formats and related options:" << endl << \
+	"	bmp" << endl << \
+	"	gif" << endl << \
+	"	jpeg 	-q <value>, --quality=<value>		Compression quality (0-95, 0 means default)." << endl << \
+	"	png	-p, --palette				Whether to use palette or not."<< endl << \
+	"		-c <value>, --compression=<value>	Compression level (0-9)." << endl << \
+	"	tiff" << endl << \
+	"	webp	-q <value>, --quality=<value>		Compression quality (0-100)." << endl
 
 #define HELP_DEMO \
-	"Available actions:" << endl << \
-	"-h, --help: print this message." << endl << endl << \
-	"demo: render a demo pfm image and write it to a PFM file (default: 'demo.pfm')." << endl << endl << \
-	"	Common options:" << endl << \
-	"		-w <value>, --width=<value>			Width of the final image." << endl << \
-	"		-h <value>, --height=<value>			Height of the final image." << endl << \
-	"		-p <string>, --projection=<string>		Projection used (default 'perspective')." <<endl << \
-	"		--angleDeg=<value>				Angle of rotation (on z axis) of the camera." <<endl << \
-	"		--antialiasing=<value>				Number of samples per single pixel (must be a perfect square, e.g. 4)." <<endl << \
-	"		-s <value>, --seed=<value>			Random number generator seed." << endl << \
-	"		-o <string>, --outfile=<string>			Filename of the output image." << endl
+	"demo: render a demo pfm image." << endl << endl << \
+	"Usage: " << programName << " demo [options] <inputfile>" << endl << endl << \
+	"General options:" << endl << \
+	"	-h, --help					Print this message." << endl << \
+	"	-w <value>, --width=<value>			Width of the final image (default 640)." << endl << \
+	"	-h <value>, --height=<value>			Height of the final image (default 480)." << endl << \
+	"	-a <value>, --aspectRatio=<value>		Aspect ratio of the final image (default width/height)." << endl << \
+	"	-p <string>, --projection=<string>		Projection used (default 'perspective'). Can be 'perspective' or 'orthogonal'" << endl << \
+	"	-D <value>, --angleDeg=<value>			Angle of rotation (on z axis) of the camera (default 0)." << endl << \
+	"	-A <value>, --antialiasing=<value>		Number of samples per single pixel (default 0). Must be a perfect square, e.g. 4." << endl << \
+	"	-R <renderer>, --renderer=<renderer>		Rendering algorithm (default 'path'). Can be 'path', 'debug', 'onoff', 'flat'." << endl << \
+	"	-o <string>, --outfile=<string>			Filename of the output image (default 'demo.pfm')." << endl << endl << \
+	"Options for 'path' rendering algorithm:" << endl << \
+	"	-s <value>, --seed=<value>			Random number generator seed (default 42)." << endl << \
+	"	-n <value>, --nRays=<value>			Number of rays started at each intersection (default 3)." << endl << \
+	"	-d <value>, --depth=<value>			Max ray depth (default 4)." << endl << \
+	"	-r <value>, --roulette=<value>			Ray depth to start Russian roulette (default 3)." << endl
 
 #define HELP_STACK \
 	"stack: stack more pfm images representing the same scene." << endl << endl << \
+	"Usage: " << programName << " stack [options] <inputfile1> [<inputfile2>] ..." << endl << endl << \
 	"Available options:" << endl << \
-	"	-h, --help: print this message." << endl << \
-	"	-m <string>, --method=<string>			The stacking method (can be 'mean' or 'median', default 'mean')." << endl << \
-	"	-S <value>, --nSigma=<value>			Number of sigma clipping iterations (default 0)." << endl << \
-	"	-a <value>, --alpha=<value>			Sigma clipping alpha factor (consider outliers values farther than alpha*sigma from the median, default 2)." << \
-	"	-o <string>, --outfile=<string>			Filename of the output image." << endl
+	"	-h, --help				Print this message." << endl << \
+	"	-m <string>, --method=<string>		The stacking method (default 'mean'). Can be 'mean' or 'median'." << endl << \
+	"	-S <value>, --nSigma=<value>		Number of sigma clipping iterations (default 0)." << endl << \
+	"	-a <value>, --alpha=<value>		Sigma clipping alpha factor (consider outliers values farther than alpha*sigma from the median, default 2)." << endl << \
+	"	-o <string>, --outfile=<string>		Filename of the output image (default 'stack.pfm')." << endl
+
+#define HELP_RENDER \
+	"render: render a scenefile to a pfm image." << endl << endl << \
+	"Usage: " << programName << " render [options] <inputfile>" << endl << endl << \
+	"General options:" << endl << \
+	"	-h, --help							Print this message." << endl << \
+	"	-f <variable1:value1,variable2:value2,...>, --float=<...>	Define float variables to be used in the scenefile." << endl << \
+	"	-w <value>, --width=<value>					Width of the final image (default 640)." << endl << \
+	"	-h <value>, --height=<value>					Height of the final image (default 480)." << endl << \
+	"	-a <value>, --aspectRatio=<value>				Aspect ratio of the final image (default width/height)." << endl << \
+	"	-A <value>, --antialiasing=<value>				Number of samples per single pixel (default 0). Must be a perfect square, e.g. 4." << endl << \
+	"	-R <renderer>, --renderer=<renderer>				Rendering algorithm (default 'path'). Can be 'path', 'debug', 'onoff', 'flat'." << endl << \
+	"	-o <string>, --outfile=<string>					Filename of output image (default input filename with '.pfm' extension)." << endl << endl <<\
+	"Options for 'path' rendering algorithm:" << endl << \
+	"	-s <value>, --seed=<value>					Random number generator seed (default 42)." << endl << \
+	"	-n <value>, --nRays=<value>					Number of rays started at each intersection (default 3)." << endl << \
+	"	-d <value>, --depth=<value>					Max ray depth (default 4)." << endl << \
+	"	-r <value>, --roulette=<value>					Ray depth to start Russian roulette (default 3)." << endl
 
 using namespace std;
 
 enum class ImageFormat { png, webp, jpeg, tiff, bmp, gif };
 
 int demo(argh::parser cmdl);
+int render(argh::parser cmdl);
 int pfm2ldr(argh::parser cmdl);
 int stackPfm(argh::parser cmdl);
+string baseFilename(string s);
 
 int main(int argc, char *argv[])
 {
 	argh::parser cmdl;
 
-	cmdl.add_params({"-a", "--afactor", "--alpha",
+	cmdl.add_params({"-a", "--afactor", "--alpha", "--aspectRatio",
 			 "-g", "--gamma",
 			 "-c", "--compression",
 			 "-q", "--quality",
+			 "-l", "--luminosity",
 			 "-w", "--width",
 			 "-h", "--height",
 			 "-p", "--projection",
-			 "--angleDeg",
-			 "--antialiasing",
+			 "-D", "--angleDeg",
+			 "-A", "--antialiasing",
+			 "-n", "--nRays",
+			 "-d", "--depth",
+			 "-r", "--roulette",
+			 "-R", "--renderer",
 			 "-o", "--outfile",
 			 "-s", "--seed",
+			 "-f", "--float", "--format",
 			 "-S", "--nSigma",
 			 "-m", "--method"});
 	cmdl.parse(argc, argv);
@@ -107,6 +146,8 @@ int main(int argc, char *argv[])
 	// Parse action
 	if (actionName == "demo") { 
 		return demo(cmdl);
+	} else if (actionName == "render") {
+		return render(cmdl);
 	} else if (actionName == "pfm2ldr") {
 		return pfm2ldr(cmdl);
 	} else if (actionName == "stack") {
@@ -115,7 +156,8 @@ int main(int argc, char *argv[])
 		cout << USAGE;
 		return 0;
 	} else {
-		cout << USAGE << endl << RUN_HELP;
+		cout << USAGE;
+		return 1;
 	}
 }
 
@@ -125,17 +167,22 @@ int pfm2ldr(argh::parser cmdl)
 	const string actionName = cmdl[1];
 
 	if (cmdl[{"-h", "--help"}]) {
-		cerr << HELP_PFM2LDR <<endl;
+		cout << HELP_PFM2LDR;
 		return 0;
 	}
 
-	if (cmdl.size() != 5) {
-		cerr << USAGE << endl << RUN_HELP;
+	if (cmdl.size() != 3) {
+		cerr << USAGE;
 		return 1;
 	}
 
+	// Parse input filename (first positional argument)
+	string infileString = cmdl[2];
+	const char *infile = infileString.c_str();
+
 	// Parse format
-	const string formatStr = cmdl[2]; 
+	string formatStr;
+	cmdl({"-f", "--format"}, "png") >> formatStr;
 	ImageFormat format;
 	if (formatStr == "png") {
 		format = ImageFormat::png;
@@ -155,10 +202,12 @@ int pfm2ldr(argh::parser cmdl)
 		return 1;
 	}
 
-	const char *infile = cmdl[3].c_str();
-	const char *outfile = cmdl[4].c_str();
+	// Parse output filename
+	string ofileString;
+	cmdl({"-o", "--outfile"}, baseFilename(infileString) + "." + formatStr) >> ofileString;
+	const char *outfile = ofileString.c_str();
 
-	// Parse parameters (also specifying default values).
+	// Parse parameters
 	float aFactor;
 	cmdl({"-a", "--afactor"}, .3f) >> aFactor;
 
@@ -173,6 +222,9 @@ int pfm2ldr(argh::parser cmdl)
 	int quality;
 	cmdl({"-q", "--quality"}, -1) >> quality;
 
+	float luminosity;
+	cmdl({"-l", "--luminosity"}, -1.f) >> luminosity;
+
 	// Read the input file
 	HdrImage img;
 	try {
@@ -183,7 +235,10 @@ int pfm2ldr(argh::parser cmdl)
 	}
 
 	// Convert HDR to LDR
-	img.normalizeImage(aFactor, .2f);
+	if (luminosity > 0.f)	// User inputted luminosity
+		img.normalizeImage(aFactor, luminosity);
+	else			// Default luminosity
+		img.normalizeImage(aFactor);
 	img.clampImage();
 
 	// Write to output file
@@ -216,37 +271,44 @@ int pfm2ldr(argh::parser cmdl)
 	return 0;
 }
 
-int demo(argh::parser cmdl) {
-
+int demo(argh::parser cmdl)
+{
+	const string programName = cmdl[0];
+	const string actionName = cmdl[1];
 	if (cmdl[{"-h", "--help"}]) {
-		cerr << HELP_DEMO <<endl;
+		cout << HELP_DEMO;
 		return 0;
 	}
+
 	int width, height;
-	cmdl({"-w", "--width"}, 1000) >> width;
-	cmdl({"-h", "--height"}, 1000) >> height;
-	float aspectRatio = (float) width / height;
+	cmdl({"-w", "--width"}, 640) >> width;
+	cmdl({"-h", "--height"}, 480) >> height;
+	float aspectRatio;
+	cmdl({"-a", "--aspectRatio"},  (float) width / height) >> aspectRatio;
+	
+	Material skyMat{DiffusiveBRDF{UniformPigment{WHITE}}, UniformPigment{WHITE}};
+	Material wallsMat{DiffusiveBRDF{CheckeredPigment{Color{.8f, .8f, .8f}, Color{.2f, .2f, .2f}, 10}}};
+	
+	Material matteGreen{DiffusiveBRDF{UniformPigment{Color{.2f, .5f, .1f}}}};
+	
+	HdrImage marbleImage{"../textures/marble_10.pfm"};
+	Material marble{DiffusiveBRDF{ImagePigment{marbleImage}}};
 
-/*
-	Material material1{SpecularBRDF(UniformPigment(Color{.8f, .8f, .8f}))};
+	HdrImage woodImage{"../textures/wood_20.pfm"};
+	Material wood{DiffusiveBRDF{ImagePigment{woodImage}}};
+	
+	HdrImage noiseImage{"../textures/noise_10.pfm"};
+	Material noise{DiffusiveBRDF{ImagePigment{noiseImage}}};
 
-	Material material2{DiffusiveBRDF(CheckeredPigment(Color{.7f, .8f, .5f}, Color{.7f, .2f, .3f}, 4))};
-
-	Material gold{SpecularBRDF(UniformPigment(Color{.8f, .8f, .2f}))};
-	Material matte{DiffusiveBRDF{UniformPigment(Color{.5f, .5f, .5f})}};
-
-	Material materialSky{DiffusiveBRDF{UniformPigment(Color{1.f, 1.f, 1.f})}, UniformPigment{Color{.6f, .6f, .8f}}};
-	Material materialGround{DiffusiveBRDF(CheckeredPigment(Color{1.f, 0.f, 0.f}, Color{1.f, 1.f, 1.f}, 2))};
-*/
-
-	string projString;
-	int angle;
-	cmdl({"-p", "--projection"}, "perspective") >> projString;
 	int seed;
 	cmdl({"-s", "--seed"}, 42) >> seed;
-	cmdl({"--angleDeg"}, 0) >> angle;
-	Transformation camTransformation{rotationY(angle*M_PI/180)*translation(Vec{-1.f, 0.f, 0.f})};
+	float angle;
+	cmdl({"-D", "--angleDeg"}, 0) >> angle;
+	Transformation camTransformation{rotationZ(angle*M_PI/180)*translation(Vec{-1.f, 0.f, 0.f})};
+
 	shared_ptr<Camera> cam;
+	string projString;
+	cmdl({"-p", "--projection"}, "perspective") >> projString;
 	if (projString == "orthogonal")
 		cam = make_shared<OrthogonalCamera>(OrthogonalCamera{aspectRatio, camTransformation});
 	else if (projString == "perspective")
@@ -259,32 +321,48 @@ int demo(argh::parser cmdl) {
 	HdrImage image{width, height};
 	World world;
 
-/*
-	world.add(Sphere{scaling(.5f)*translation(Vec{1.2f, -1.2f, 0.f}), material1});
-	world.add(Sphere{scaling(.5f)*translation(Vec{0.f, .5f, 0.f}), material2});
-	//world.add(CSGUnion{Box{Point{-.5f, -.5f, 0.f}, Point{.5f, .5f, 1.f}, matte}, Sphere{translation(Vec{0.f, 0.f, 1.f}), gold}});
-	world.add(Sphere{scaling(5.f), materialSky});
-	world.add(Plane{translation(Vec{0.f, 0.f, -1.f}), materialGround});
-*/
+	world.add(Sphere{translation(Vec{3.5f, 2.f, 0.f}), marble});
+	world.add(Sphere{translation(Vec{3.5f, 0.f, 0.f}), wood});
+	world.add(Sphere{translation(Vec{3.5f, -2.f, 0.f}), noise});
+	world.add(Sphere{translation(Vec{1.5f, 0.f, -2.4f})*scaling(.5f), skyMat});
 
-	world.add(Sphere{});
-	//world.add(Sphere{scaling(.5f)*translation(Vec{0.f, .5f, 0.f})});
-	world.add(Plane{translation(Vec{0.f, 0.f, 4.f})});
-	world.add(Plane{translation(Vec{0.f, 0.f, -4.f})});
-	
+
+	world.add(Box{Point{-1.5f, -3.5f, -3.5f}, Point{5.5f, 3.5f, 3.5f}, wallsMat});
+	world.add(Box{Point{-1.5f, -3.5f, 3.4f}, Point{5.5f, 3.5f, 3.5f}, skyMat});
+	world.add(Box{Point{-1.5f, -3.5f, -3.5f}, Point{5.5f, 3.5f, -3.4f}, matteGreen});
 
 	int samplesPerPixel;
-	cmdl({"--antialiasing"}, 0) >> samplesPerPixel;
+	cmdl({"-A", "--antialiasing"}, 0) >> samplesPerPixel;
 	int samplesPerSide = sqrt(samplesPerPixel);
 	if (samplesPerPixel != samplesPerSide*samplesPerSide){
 		cerr << "Not a perfect square given as --antialiasing parameter."  <<endl;
 		return 1;
-	} 
+	}
 	ImageTracer tracer{image, *cam, samplesPerSide};
 	PCG pcg{(uint64_t) seed};
 
-	//tracer.fireAllRays(PathTracer{world, pcg, 2, 4, 5});
-	tracer.fireAllRays(DebugRenderer(world));
+	int nRays;
+	cmdl({"-n", "--nRays"}, 3) >> nRays;
+	int depth;
+	cmdl({"-d", "--depth"}, 4) >> depth;
+	int roulette;
+	cmdl({"-r", "--roulette"}, 3) >> roulette;
+
+	string renderer;
+	cmdl({"-R", "--renderer"}, "path") >> renderer;
+
+	if (renderer == "path")
+		tracer.fireAllRays(PathTracer{world, pcg, nRays, depth, roulette}, true);
+	else if (renderer == "debug")
+		tracer.fireAllRays(DebugRenderer{world});
+	else if (renderer == "onoff")
+		tracer.fireAllRays(OnOffRenderer{world});
+	else if (renderer == "flat")
+		tracer.fireAllRays(FlatRenderer{world});
+	else {
+		cerr << "Error: renderer " << renderer << " not supported" << endl;
+		return 1;
+	}
 
 	string ofilename;
 	cmdl({"-o", "--outfile"}, "demo.pfm") >> ofilename;
@@ -297,12 +375,121 @@ int demo(argh::parser cmdl) {
 
 }
 
-int stackPfm(argh::parser cmdl)
+int render(argh::parser cmdl)
 {
 	const string programName = cmdl[0];
 	const string actionName = cmdl[1];
 	if (cmdl[{"-h", "--help"}]) {
-		cerr << HELP_STACK << endl;
+		cout << HELP_RENDER;
+		return 0;
+	}
+
+	if (cmdl.size() != 3) {
+		cerr << USAGE << endl << HELP_RENDER;
+		return 1;
+	}
+
+	int width, height;
+	cmdl({"-w", "--width"}, 640) >> width;
+	cmdl({"-h", "--height"}, 480) >> height;
+	float aspectRatio;
+	cmdl({"-a", "--aspectRatio"},  (float) width / height) >> aspectRatio;
+
+	int seed;
+	cmdl({"-s", "--seed"}, 42) >> seed;
+
+	int samplesPerPixel;
+	cmdl({"-A", "--antialiasing"}, 0) >> samplesPerPixel;
+	int samplesPerSide = sqrt(samplesPerPixel);
+	if (samplesPerPixel != samplesPerSide*samplesPerSide){
+		cerr << "Not a perfect square given as --antialiasing parameter."  <<endl;
+		return 1;
+	} 
+
+	string ifilename = cmdl[2];
+	ifstream ifile{ifilename};
+	if (!ifile.is_open()) {
+		cerr << "Cannot open file " + cmdl[2] << endl;
+		return 1;
+	}
+	InputStream input{ifile, cmdl[2]};
+
+	int nRays;
+	cmdl({"-n", "--nRays"}, 3) >> nRays;
+	int depth;
+	cmdl({"-d", "--depth"}, 4) >> depth;
+	int roulette;
+	cmdl({"-r", "--roulette"}, 3) >> roulette;
+
+	string renderer;
+	cmdl({"-R", "--renderer"}, "path") >> renderer;
+
+	string variablesString;
+	cmdl({"-f", "--float"}, string{}) >> variablesString;
+	stringstream variablesStream{variablesString};
+	variablesStream.peek();	// To set eofbit if the stream is empty.
+	unordered_map<string, float> variables;
+	while (!variablesStream.eof()) {
+		string name;
+		getline(variablesStream, name, ':');
+		if (name.empty() or variablesStream.fail()) {
+			cerr << "Error: expected variable name in --float definition" << endl;
+			return 1;
+		} else if (variablesStream.eof()) {
+			cerr << "Error: expected : after variable name in --float definition" << endl;
+			return 1;
+		}
+		float value;
+		variablesStream >> value;
+		if (variablesStream.fail()) {
+			cerr << "Error: expected variable value after : in --float definition" << endl;
+			return 1;
+		}
+		variables.insert({name, value});
+		if (!variablesStream.eof() and variablesStream.get() != ',') {
+			cerr << "Error: expected , or string end after value in --float definition" << endl;
+			return 1;
+		}
+	}
+	try {
+		Scene scene{input.parseScene(variables, aspectRatio)};
+		HdrImage image{width, height};
+		ImageTracer tracer{image, *scene.camera, samplesPerSide};
+		PCG pcg{(uint64_t) seed};
+		if (renderer == "path")
+			tracer.fireAllRays(PathTracer{scene.world, pcg, nRays, depth, roulette}, true);
+		else if (renderer == "debug")
+			tracer.fireAllRays(DebugRenderer{scene.world});
+		else if (renderer == "onoff")
+			tracer.fireAllRays(OnOffRenderer{scene.world});
+		else if (renderer == "flat")
+			tracer.fireAllRays(FlatRenderer{scene.world});
+		else {
+			cerr << "Error: renderer " << renderer << " not supported" << endl;
+			return 1;
+		}
+
+		string ofilename;
+		cmdl({"-o", "--outfile"}, baseFilename(ifilename) + ".pfm") >> ofilename;
+		ofstream outPfm;
+		outPfm.open(ofilename);
+		image.writePfm(outPfm);
+		outPfm.close();
+	} catch (GrammarError &e) {
+		cerr << "Error: " << string{e.location} << ": " << e.what() << endl;
+		return 1;
+	}
+
+	return 0;
+}
+
+int stackPfm(argh::parser cmdl)
+{
+	const string programName = cmdl[0];
+	const string actionName = cmdl[1];
+
+	if (cmdl[{"-h", "--help"}]) {
+		cout << HELP_STACK;
 		return 0;
 	}
 
@@ -432,4 +619,13 @@ int stackPfm(argh::parser cmdl)
 	outPfm.close();
 
 	return 0;
+}
+
+// Return the filename without the path and the extension
+// E.g.: baseFilename("/usr/include/stdio.h") == "stdio".
+string baseFilename(string s)
+{
+	s = s.substr(s.find_last_of('/')+1);
+	s = s.substr(0, s.find_last_of('.'));
+	return s;
 }